backend/requirements.txt
Flask
SQLAlchemy
requests
python-dotenv
pyorbital
google-api-python-client
earthengine-api
celery